+++ main.ts
@@ -1,9 +1,7 @@
-import { wrap } from "module";
-import { Editor, MarkdownView, Plugin, Setting, PluginSettingTab, App, Menu, ButtonComponent, requestUrl, EditorPosition, Notice } from "obsidian";
-import { decode } from "querystring";
-import { array2markdown, table2bullet, capitalizeWord, capitalizeSentence, removeAllSpaces, zoteroNote, textWrapper, replaceLigature, ankiSelection, sortTodo, requestAPI, headingLevel, slugify, snakify, extraDoubleSpaces, toTitleCase } from "src/format";
+import { Editor, MarkdownView, Plugin, Notice } from "obsidian";
 import { removeWikiLink, removeUrlLink, url2WikiLink } from "src/link";
 import { FormatSettings, DEFAULT_SETTINGS, TextFormatSettingTab } from "src/setting";
+import { array2markdown, table2bullet, capitalizeWord, capitalizeSentence, removeAllSpaces, zoteroNote, textWrapper, replaceLigature, ankiSelection, sortTodo, requestAPI, headingLevel, slugify, snakify, extraDoubleSpaces, toTitleCase } from "src/format";
 
 export default class TextFormat extends Plugin {
   settings: FormatSettings;
@@ -12,9 +10,12 @@ export default class TextFormat extends Plugin {
     await this.loadSettings();
     this.addSettingTab(new TextFormatSettingTab(this.app, this));
 
+    let lang = window.localStorage.getItem('language');
+    if (["en", "zh", "zh-TW"].indexOf(lang) == -1) { lang = "en"; }
+
     this.addCommand({
       id: "text-format-heading-upper",
-      name: "Heading upper",
+      name: { "en": "Heading upper", "zh": "标题上升一级", "zh-TW": "標題上升一級" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "heading", true);
       },
@@ -23,12 +24,11 @@ export default class TextFormat extends Plugin {
         {
           modifiers: ["Ctrl", "Shift"],
           key: "]",
-        },
-      ],
+        }],
     });
     this.addCommand({
       id: "text-format-heading-lower",
-      name: "Heading lower",
+      name: { "en": "Heading lower", "zh": "标题下降一级", "zh-TW": "標題下降一級" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "heading", false);
       },
@@ -37,13 +37,12 @@ export default class TextFormat extends Plugin {
         {
           modifiers: ["Ctrl", "Shift"],
           key: "[",
-        },
-      ],
+        }],
     });
     this.settings.WrapperList.forEach((wrapper, index) => {
       this.addCommand({
         id: `text-format-wrapper-${index}`,
-        name: "Wrapper - " + wrapper.name,
+        name: { "en": "Wrapper", "zh": "包装器", "zh-TW": "包裝器" }[lang] + " - " + wrapper.name,
         editorCallback: (editor: Editor, view: MarkdownView) => {
           textWrapper(editor, view, wrapper.prefix, wrapper.suffix)
         },
@@ -52,7 +51,7 @@ export default class TextFormat extends Plugin {
     this.settings.RequestList.forEach((request, index) => {
       this.addCommand({
         id: `text-format-request-${index}`,
-        name: "API Request - " + request.name,
+        name: { "en": "API Request", "zh": "API 请求", "zh-TW": "API 請求" }[lang] + " - " + request.name,
         editorCallback: (editor: Editor, view: MarkdownView) => {
           this.textFormat(editor, view, "api-request", request.url);
         },
@@ -60,262 +59,253 @@ export default class TextFormat extends Plugin {
     });
     this.addCommand({
       id: "text-format-anki-card",
-      name: "Convert selection into Anki card format",
+      name: { "en": "Convert selection into Anki card format", "zh": "将选中内容转换为 Anki 卡片格式", "zh-TW": "將選取內容轉換為 Anki 卡片格式" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "anki");
       },
     });
     this.addCommand({
       id: "text-format-ligature",
-      name: "Replace ligature",
+      name: { "en": "Replace ligature", "zh": "替换连字", "zh-TW": "取代連字" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "ligature");
       },
     });
     this.addCommand({
       id: "text-format-remove-wiki-link",
-      name: "Remove WikiLinks format in selection",
+      name: { "en": "Remove WikiLinks format in selection", "zh": "移除选中文本中的 WikiLinks 格式", "zh-TW": "移除選取文字中的 WikiLinks 格式" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "remove-wiki-link");
       },
     });
     this.addCommand({
       id: "text-format-remove-url-link",
-      name: "Remove URL links format in selection",
+      name: { "en": "Remove URL links format in selection", "zh": "移除选中文本中的 URL 链接格式", "zh-TW": "移除選取文字中的 URL 鏈接格式" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "remove-url-link");
       },
     });
     this.addCommand({
       id: "text-format-link-url2wiki",
-      name: "Convert URL links to WikiLinks in selection",
+      name: { en: "Convert URL links to WikiLinks in selection", zh: "将选中文本中的 URL 链接转换为 WikiLinks 格式", "zh-TW": "將選取文字中的 URL 鏈接轉換為 WikiLinks 格式" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "link-url2wiki");
       },
     });
     this.addCommand({
       id: "text-format-lower",
-      name: "Lowercase selected text",
+      name: { en: "Lowercase selected text", zh: "将选中文本转换为小写", "zh-TW": "將選取文字轉換為小寫" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "lowercase");
       },
     });
     this.addCommand({
       id: "text-format-upper",
-      name: "Uppercase selected text",
+      name: { en: "Uppercase selected text", zh: "将选中文本转换为大写", "zh-TW": "將選取文字轉換為大寫" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "uppercase");
       },
     });
     this.addCommand({
       id: "text-format-togglecase",
-      name: "Togglecase selected text",
+      name: { en: "Togglecase selected text", zh: "触发选中文本大小写切换", "zh-TW": "觸發選取文字大小寫切換" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "togglecase");
       },
     });
     this.addCommand({
       id: "text-format-capitalize-word",
-      name: "Capitalize all words in selected text",
+      name: { en: "Capitalize all words in selected text", zh: "将选中文本中的所有单词首字母大写", "zh-TW": "將選取文字中的所有單詞首字母大寫" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "capitalize-word");
       },
     });
     this.addCommand({
       id: "text-format-capitalize-sentence",
-      name: "Capitalize only first word of sentence in selected text",
+      name: { en: "Capitalize only first word of sentence in selected text", zh: "将选中文本中的句子的首字母大写", "zh-TW": "將選取文字中的句子的首字母大寫" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "capitalize-sentence");
       },
     });
     this.addCommand({
       id: "text-format-titlecase",
-      name: "Title case selected text",
+      name: { en: "Title case selected text", zh: "将选中文本转换为标题格式大小写", "zh-TW": "將選取文字轉換為標題格式大小寫" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "titlecase");
       },
     });
     this.addCommand({
       id: "text-format-remove-spaces",
-      name: "Remove redundant spaces in selection",
+      name: { en: "Remove redundant spaces in selection", zh: "将选中文本中的多余空格移除", "zh-TW": "將選取文字中的多餘空格移除" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "remove-spaces");
       },
     });
     this.addCommand({
       id: "text-format-remove-spaces-all",
-      name: "Remove all spaces in selection",
+      name: { en: "Remove all spaces in selection", zh: "将选中文本中的所有空格移除", "zh-TW": "將選取文字中的所有空格移除" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "spaces-all");
       },
     });
     this.addCommand({
       id: "text-format-remove-blank-line",
-      name: "Remove blank line(s)",
+      name: { en: "Remove blank line(s)", zh: "将选中文本中的空行移除", "zh-TW": "將選取文字中的空行移除" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "remove-blank-line");
       },
     });
     this.addCommand({
       id: "text-format-merge-line",
-      name: "Merge broken paragraph(s) in selection",
+      name: { en: "Merge broken paragraph(s) in selection", zh: "将选中文本中的断行合并", "zh-TW": "將選取文字中的斷行合併" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "merge");
       },
     });
     this.addCommand({
       id: "text-format-bullet-list",
-      name: "Format bullet list",
+      name: { en: "Detect and format bullet list", zh: "识别并格式化无序列表", "zh-TW": "識別並格式化無序清單" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "bullet");
       },
     });
     this.addCommand({
       id: "text-format-convert-ordered-list",
-      name: "Format ordered list",
+      name: { en: "Detect and format ordered list", zh: "识别并格式化有序列表", "zh-TW": "識別並格式化有序清單" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "convert-ordered");
       },
     });
     this.addCommand({
       id: "text-format-split-blank",
-      name: "Split line(s) by blanks",
+      name: { en: "Split line(s) by blanks", zh: "将选中文本按空格分行", "zh-TW": "將選取文字按空格分行" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "split-blank");
       },
     });
     this.addCommand({
       id: "text-format-chinese-punctuation",
-      name: "Convert to Chinese punctuation marks (,;:!?)",
+      name: { en: "Convert to Chinese punctuation marks (,;:!?)", zh: "转换为中文标点符号（,;:!?）", "zh-TW": "轉換為中文標點符號（,;:!?）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "Chinese-punctuation");
       },
     });
     this.addCommand({
       id: "text-format-english-punctuation",
-      name: "Convert to English punctuation marks",
+      name: { en: "Convert to English punctuation marks", zh: "转换为英文标点符号", "zh-TW": "轉換為英文標點符號" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "English-punctuation");
       },
     });
     this.addCommand({
       id: "text-format-latex-single-letter",
-      name: "Convert single letter into math mode (latex)",
+      name: { en: "Convert single letter into math mode (LaTeX)", zh: "将单个字母转换为数学模式（LaTeX）", "zh-TW": "將單個字母轉換為數學模式（LaTeX）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "latex-letter");
       },
     });
     this.addCommand({
       id: "text-format-decodeURI",
-      name: "Decode URL",
+      name: { en: "Decode URL", zh: "解码 URL", "zh-TW": "解碼 URL" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "decodeURI");
       },
     });
     this.addCommand({
       id: "text-format-paragraph-double-spaces",
-      name: "Add extra double spaces per paragraph for whole file",
+      name: { en: "Add extra double spaces per paragraph for whole file", zh: "全文为每段段末添加双空格", "zh-TW": "全文為每段段末添加雙空格" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         extraDoubleSpaces(editor, view);
       },
     });
     this.addCommand({
       id: "text-format-add-line-break",
-      name: "Add extra line break to paragraph",
+      name: { en: "Add extra line break to paragraph", zh: "在段落末添加额外换行", "zh-TW": "在段落末添加額外換行" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "add-line-break");
       },
     });
     this.addCommand({
       id: "text-format-hyphen",
-      name: "Remove hyphens",
+      name: { en: "Remove hyphens", zh: "移除连字符", "zh-TW": "移除連字符" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "hyphen");
       },
     });
     this.addCommand({
       id: "text-format-remove-citation-index",
-      name: "Remove citation index",
+      name: { en: "Remove citation index", zh: "移除引用索引编号", "zh-TW": "移除引用索引編號" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "remove-citation");
       },
     });
     this.addCommand({
       id: "text-format-mathpix-array2table",
-      name: "Convert Mathpix array to markdown table",
+      name: {
+        en: "Convert Mathpix's LaTeX array to markdown table", zh: "将 Mathpix 的 LaTeX 数组转换为 Markdown 表格", "zh-TW": "將 Mathpix 的 LaTeX 陣列轉換為 Markdown 表格"
+      }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "array2table");
       },
     });
     this.addCommand({
       id: "text-format-table2bullet",
-      name: "Convert table to bullet list without header",
+      name: { en: "Convert table to bullet list without header", zh: "将表格转换为无序列表（不含标题）", "zh-TW": "將表格轉換為無序清單（不含標題）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "table2bullet");
       },
     });
     this.addCommand({
       id: "text-format-table2bullet-head",
-      name: "Convert table to bullet list with header",
+      name: { en: "Convert table to bullet list with header", zh: "将表格转换为无序列表（含标题）", "zh-TW": "將表格轉換為無序清單（含標題）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "table2bullet-header");
       },
     });
     this.addCommand({
       id: "text-format-todo-sort",
-      name: "Sort to-do list",
+      name: { en: "Sort to-do list", zh: "将待办事项列表排序", "zh-TW": "將待辦事項列表排序" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "todo-sort");
       },
     });
     this.addCommand({
       id: "text-format-slugify",
-      name: "Slugify selected text",
+      name: { en: "Slugify selected text (`-` for space)", zh: "使用 Slugify 格式化选中文本（`-`连字符）", "zh-TW": "使用 Slugify 格式化選取文字（`-`連字符）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "slugify");
       },
     })
     this.addCommand({
       id: "text-format-snakify",
-      name: "Snakify selected text",
+      name: { en: "Snakify selected text (`_` for space)", zh: "使用 Snakify 格式化选中文本（`_`连字符）", "zh-TW": "使用 Snakify 格式化選取文字（`_`連字符）" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "snakify");
       },
     })
     this.addCommand({
       id: "text-format-space-word-symbol",
-      name: "Format space between word and symbol",
+      name: { en: "Format space between word and symbol", zh: "格式化单词与符号之间的空格", "zh-TW": "格式化單詞與符號之間的空格" }[lang],
       editorCallback: (editor: Editor, view: MarkdownView) => {
         this.textFormat(editor, view, "space-word-symbol");
       },
     });
     this.addCommand({
       id: "text-format-zotero-note",
-      name: "Zotero note format and paste",
-      callback: async () => {
+      name: { en: "Get Zotero note from clipboard and paste", zh: "从剪贴板获取 Zotero 笔记并粘贴", "zh-TW": "從剪貼板獲取 Zotero 筆記並粘貼" }[lang],
+      editorCallback: async (editor: Editor, view: MarkdownView) => {
         const clipboardText = await navigator.clipboard.readText();
         let text = zoteroNote(
           clipboardText,
           this.settings.ZoteroNoteRegExp,
           this.settings.ZoteroNoteTemplate
         );
-        let markdownView = this.app.workspace.getActiveViewOfType(MarkdownView);
-        if (!markdownView) {
-          return;
-        }
-        let editor = markdownView.editor;
         editor.replaceSelection(text);
       },
     });
   }
 
   textFormat(editor: Editor, view: MarkdownView, cmd: string, args: any = ""): void {
-    // let markdownView = this.app.workspace.getActiveViewOfType(MarkdownView);
-    if (!view) {
-      return;
-    }
-    // let editor = markdownView.editor;
-
     var selectedText: string, replacedText;
 
     // if nothing is selected, select the whole line.