+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 - [feature] New Setting: Proper nouns, which are ignored in title case. e.g., USA, UFO.
+- [updated] paste Zotero note: if regex fail to match, return original text in clipboard.
+- [updated] General: commands support Chinese language. 命令支持中文显示
 
 ## 2.4.1
 - [fix] support canvas
@@ -62,7 +64,7 @@
 - [fix] #38
 
 ## 2.2.0
-- [merged] toggle-case
+- [feature] merge: toggle-case
 
 ## 2.1.0
 - [feature] API request